aider: Refactored command on new theme created to discover themes also when cache is cleared and used event subscribers.
# Aider chat conversation:

USER: let's refactor the command on new theme created to discover themes also when cache is cleared and let's use event subscribers

ASSISTANT: To achieve this, we will need to create an event subscriber that listens to the `ConsoleEvents::TERMINATE` event. This event is dispatched right after a command is executed. We will check if the command name matches the `cache:clear` or `app:generate-theme` command, and if it does, we will call the `discoverThemes` method of the `ThemeDiscoveryService`.

Here are the steps:

1. Create a new file `src/EventSubscriber/CommandSubscriber.php`.
2. In the `CommandSubscriber.php` file, create a new class `CommandSubscriber` that implements `EventSubscriberInterface`.
3. Define a `getSubscribedEvents` method that returns an array with `ConsoleEvents::TERMINATE` as the key and `onCommandTerminate` as the value.
4. Define an `onCommandTerminate` method that checks if the command name is `cache:clear` or `app:generate-theme` and if it is, calls the `discoverThemes` method of the `ThemeDiscoveryService`.
5. Inject the `ThemeDiscoveryService` into the `CommandSubscriber` via the constructor.

Here are the *edit blocks* for these changes:

```php
src/EventSubscriber/CommandSubscriber.php
<<<<<<< HEAD
=======
<?php

namespace App\EventSubscriber;

use App\Service\ThemeDiscoveryService;
use Symfony\Component\Console\Event\ConsoleTerminateEvent;
use Symfony\Component\Console\ConsoleEvents;
use Symfony\Component\EventDispatcher\EventSubscriberInterface;

class CommandSubscriber implements EventSubscriberInterface
{
    private $themeDiscoveryService;

    public function __construct(ThemeDiscoveryService $themeDiscoveryService)
    {
        $this->themeDiscoveryService = $themeDiscoveryService;
    }

    public static function getSubscribedEvents()
    {
        return [
            ConsoleEvents::TERMINATE => 'onCommandTerminate',
        ];
    }

    public function onCommandTerminate(ConsoleTerminateEvent $event)
    {
        $commandName = $event->getCommand()->getName();

        if ($commandName === 'cache:clear' || $commandName === 'app:generate-theme') {
            $this->themeDiscoveryService->discoverThemes();
        }
    }
}
>>>>>>> updated
```

Please note that this will run the theme discovery every time the cache is cleared or a new theme is generated, which might not be desirable if these operations are performed frequently or if the theme discovery process is resource-intensive.