aider: Adjusted approach to create a new `Theme` instance every time a theme is discovered, regardless of whether it already exists or not.
# Aider chat conversation:

USER: theme entity doesn't have this property since it's immutable $theme->setDescription

ASSISTANT: I see, if the `Theme` entity is immutable and doesn't have a `setDescription` method, we need to adjust our approach. Instead of updating an existing theme, we should create a new `Theme` instance every time we discover a theme. This way, we ensure that the theme's description and author are always up-to-date.

Here's the *edit block* for the `discoverThemes` method in the `ThemeDiscoveryService`:

```php
src/Service/ThemeDiscoveryService.php
<<<<<<< HEAD
            if ($this->filesystem->exists($composerJsonPath)) {
                $composerJson = json_decode(file_get_contents($composerJsonPath), true, 512, JSON_THROW_ON_ERROR);

                if (!$theme) {
                    $theme = new Theme($themeName, $composerJson['description'] ?? '', $composerJson['authors'][0]['name'] ?? '');
                    $this->entityManager->persist($theme);
                }
                $theme->setDescription($composerJson['description'] ?? '');
                $theme->setAuthor($composerJson['authors'][0]['name'] ?? '');
            }
=======
            if ($this->filesystem->exists($composerJsonPath)) {
                $composerJson = json_decode(file_get_contents($composerJsonPath), true, 512, JSON_THROW_ON_ERROR);

                $theme = new Theme($themeName, $composerJson['description'] ?? '', $composerJson['authors'][0]['name'] ?? '');
                $this->entityManager->persist($theme);
            }
>>>>>>> updated
```

This change will ensure that a new `Theme` instance is created every time a theme is discovered, regardless of whether the theme already exists or not.