aider: Added optional `id` parameter to `createTheme` method in `ThemeManager` class.
# Aider chat conversation:

USER: you forgot to pass parameter id which by default is null in case it's not a theme update

ASSISTANT: I see, you want to add an optional `id` parameter to the `createTheme` method in the `ThemeManager` class. This `id` will be `null` by default, indicating that a new theme is being created. If an `id` is provided, it means we are updating an existing theme.

Here's the *edit block* for the changes in `src/Service/ThemeManager.php`:

```php
src/Service/ThemeManager.php
<<<<<<< HEAD
    public function createTheme(string $name, string $title, string $description, string $author, bool $isActive): Theme
    {
        $theme = new Theme($name, $title, $description, $author, new \DateTime(), new \DateTime(), $isActive);
        $this->entityManager->persist($theme);
        $this->entityManager->flush();

        return $theme;
    }
=======
    public function createTheme(string $name, string $title, string $description, string $author, bool $isActive, ?int $id = null): Theme
    {
        $theme = new Theme($id, $name, $title, $description, $author, new \DateTime(), new \DateTime(), $isActive);
        $this->entityManager->persist($theme);
        $this->entityManager->flush();

        return $theme;
    }
>>>>>>> updated
```

In this updated version, the `createTheme` method now accepts an optional `id` parameter. When creating a new `Theme` object, this `id` is passed to the constructor.