aider: Refactored `themes.html.twig` to not use getters and setters for the theme.
# Aider chat conversation:

USER: refactor themes.html.twig to not use getters and setters for the theme

ASSISTANT: To refactor the `themes.html.twig` to not use getters and setters for the theme, we need to modify the `themes.html.twig` file.

Here are the steps:

1. In the `themes.html.twig` file, we need to change the way we access the properties of the theme. Instead of calling getters like `theme.getTitle()`, `theme.getDescription()`, `theme.getAuthor()`, `theme.getCreatedAt()`, `theme.getUpdatedAt()`, we will directly access the properties like `theme.title`, `theme.description`, `theme.author`, `theme.createdAt`, `theme.updatedAt`.

Here are the changes:

```twig
templates/themes.html.twig
<<<<<<< HEAD
            <td>{{ theme.getTitle() }}</td>
            <td>{{ theme.getDescription() }}</td>
            <td>{{ theme.getAuthor() }}</td>
            <td>{{ theme.getCreatedAt()|date('Y-m-d H:i') }}</td>
            <td>{{ theme.getUpdatedAt()|date('Y-m-d H:i') }}</td>
=======
            <td>{{ theme.title }}</td>
            <td>{{ theme.description }}</td>
            <td>{{ theme.author }}</td>
            <td>{{ theme.createdAt|date('Y-m-d H:i') }}</td>
            <td>{{ theme.updatedAt|date('Y-m-d H:i') }}</td>
>>>>>>> updated
```

This change ensures that we are not using getters and setters to access the properties of the theme in the `themes.html.twig` file.