aider: Added a question to ask for the version of the theme and included it in the composer.json file when generating a new theme.
# Aider chat conversation:

USER: proceed with updating the files but do not edit composer.json

ASSISTANT: Alright, let's proceed with the changes.

Next, we need to update `src/Command/GenerateThemeCommand.php` to include the `version` when generating a new theme.

Here is the *edit block* for this change:

```php
src/Command/GenerateThemeCommand.php
<<<<<<< HEAD
        $composerJson = [
            "name" => $packageName,
            "title" => $themeTitle,
            "description" => $description,
            "license" => $license,
            "authors" => [
                [
                    "name" => $author,
                    "email" => $email
                ]
            ]
        ];
=======
        $question = new Question('Please enter the version of the theme: ');
        $version = $helper->ask($input, $output, $question);

        $composerJson = [
            "name" => $packageName,
            "title" => $themeTitle,
            "description" => $description,
            "license" => $license,
            "version" => $version,
            "authors" => [
                [
                    "name" => $author,
                    "email" => $email
                ]
            ]
        ];
>>>>>>> updated
```

This *edit block* adds a question to ask for the `version` of the theme and includes it in the `composer.json` file when generating a new theme.