aider: Implemented the `ToEntity` trait which creates a new instance of an entity and populates it with data from the DTO.
# Aider chat conversation:

USER: let's also implement a trait called ToEntity which would return a new instance of an entity populated with the data from the dto

ASSISTANT: To implement the `ToEntity` trait, we need to create a new file `src/Domain/Traits/ToEntity.php`. However, the implementation of this trait will depend on the specific entity that it will be used with.

Could you please provide more details about the entity that this trait will be used with? Specifically, I need to know the name of the entity and its properties.

USER: the entity should be passed as an argument for example Theme::class from `src/Entity/Theme.php`

ASSISTANT: To implement the `ToEntity` trait, we need to create a new file `src/Domain/Traits/ToEntity.php` and add the trait code to it.

Here is the change we need to make:

```php
src/Domain/Traits/ToEntity.php
<<<<<<< HEAD

=======
<?php

namespace App\Domain\Traits;

trait ToEntity
{
    public function toEntity(string $entityClass)
    {
        $entity = new $entityClass();

        foreach ($this as $property => $value) {
            if (property_exists($entity, $property)) {
                $entity->$property = $value;
            }
        }

        return $entity;
    }
}
>>>>>>> updated
```

This change creates a new trait `ToEntity` in the `App\Domain\Traits` namespace. The trait has a method `toEntity` that takes a class name as an argument, creates a new instance of that class, and assigns the properties of the current object to the new entity if they exist. The method then returns the new entity.