aider: Modified the validation for the package name in `GenerateThemeCommand.php` to allow for the format `vendor/package`.
# Aider chat conversation:

USER: the validation for composer valid package is not ok it will not allow me to enter a package name ex zenchron/nexus-theme doctrine/doctrine-bundle symfony/console

ASSISTANT: I see, you're right. The current validation only allows for a single word package name, but composer package names usually consist of a vendor name and a package name separated by a slash.

Let's modify the validation for the package name in the `GenerateThemeCommand.php` file to allow for this format.

Here is the change:

```php
src/Command/GenerateThemeCommand.php
<<<<<<< HEAD
        $question = new Question('Please enter the package name of the theme: ');
        $question->setValidator(function ($answer) {
            if (!preg_match('/^[a-z0-9]+(?:-[a-z0-9]+)*$/', $answer)) {
                throw new \RuntimeException('The package name should be a valid composer package name.');
            }

            return strtolower($answer);
        });
=======
        $question = new Question('Please enter the package name of the theme: ');
        $question->setValidator(function ($answer) {
            if (!preg_match('/^[a-z0-9]+(?:-[a-z0-9]+)*\/[a-z0-9]+(?:-[a-z0-9]+)*$/', $answer)) {
                throw new \RuntimeException('The package name should be a valid composer package name.');
            }

            return strtolower($answer);
        });
>>>>>>> updated
```

This change will ensure that the package name validation allows for a format of `vendor/package`.