aider: Fixed incorrect arguments passed to `ConsoleTerminateEvent` constructor by retrieving the command instance from `CommandTester` object and passing it as the first argument.
# Aider chat conversation:

USER: we have an error TypeError: Symfony\Component\Console\Event\ConsoleTerminateEvent::__construct(): Argument #1 ($command) must be of type Symfony\Component\Console\Command\Command, Symfony\Component\Console\Input\ArrayInput given, called in /app/tests/Command/GenerateThemeCommandTest.php on line 41

ASSISTANT: The error is due to the incorrect arguments passed to the `ConsoleTerminateEvent` constructor. The first argument should be an instance of `Symfony\Component\Console\Command\Command`, but an instance of `Symfony\Component\Console\Input\ArrayInput` is being passed instead.

To fix this, we need to pass the command instance to the `ConsoleTerminateEvent` constructor. We can get the command instance from the `CommandTester` object using the `getCommand` method.

Here is the proposed change:

```php
tests/Command/GenerateThemeCommandTest.php
<<<<<<< HEAD
        $this->commandTester->execute([]);

        $output = $this->commandTester->getDisplay();
        $exitCode = $this->commandTester->getStatusCode();

        $this->eventDispatcher->dispatch(new ConsoleTerminateEvent($this->commandTester->getInput(), $exitCode));
=======
        $command = $this->commandTester->getCommand();
        $this->commandTester->execute([]);

        $output = $this->commandTester->getDisplay();
        $exitCode = $this->commandTester->getStatusCode();

        $this->eventDispatcher->dispatch(new ConsoleTerminateEvent($command, $this->commandTester->getInput(), $exitCode));
>>>>>>> updated
```

This change retrieves the command instance from the `CommandTester` object and passes it to the `ConsoleTerminateEvent` constructor.