aider: Pass `$this->commandTester->getOutput()` instead of `$exitCode` in the `ConsoleTerminateEvent` constructor in `tests/Command/GenerateThemeCommandTest.php`.
# Aider chat conversation:

ASSISTANT: I see, the `ConsoleTerminateEvent` constructor expects an instance of `OutputInterface` as the third argument, but an integer is being provided. We need to pass `$this->commandTester->getOutput()` instead of `$exitCode`.

Here is the change you need to make in the `tests/Command/GenerateThemeCommandTest.php` file:

```php
tests/Command/GenerateThemeCommandTest.php
<<<<<<< HEAD
        $this->eventDispatcher->dispatch(new ConsoleTerminateEvent($this->command, $this->commandTester->getInput(), $exitCode));
=======
        $this->eventDispatcher->dispatch(new ConsoleTerminateEvent($this->command, $this->commandTester->getInput(), $this->commandTester->getOutput()));
>>>>>>> updated
```

Please make this change and let me know if you encounter any further issues.