aider: Add $this->dispatcher to the GenerateThemeCommand class constructor.
# Aider chat conversation:

USER: $this->dispatcher is missing from the command

USER: In `src/Command/GenerateThemeCommand.php` $this->dispatcher is not defined

ASSISTANT: To use the `EventDispatcherInterface` in the `GenerateThemeCommand` class, you need to inject it into the constructor and assign it to a private property. Here is the *edit block* for this change:

```php
src/Command/GenerateThemeCommand.php
<<<<<<< HEAD
class GenerateThemeCommand extends Command
{
    private GenerateThemeCommandValidator $validator;

    public function __construct(GenerateThemeCommandValidator $validator)
    {
        $this->validator = $validator;

        parent::__construct();
    }
=======
use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;

class GenerateThemeCommand extends Command
{
    private GenerateThemeCommandValidator $validator;
    private EventDispatcherInterface $dispatcher;

    public function __construct(GenerateThemeCommandValidator $validator, EventDispatcherInterface $dispatcher)
    {
        $this->validator = $validator;
        $this->dispatcher = $dispatcher;

        parent::__construct();
    }
>>>>>>> updated
```

This will allow you to use `$this->dispatcher->dispatch($event);` in your code.