aider: Made `$parentTheme` property public in `ThemeData` class.
# Aider chat conversation:

USER: ThemeData should have public for $parentTheme

ASSISTANT: To make the `$parentTheme` property public in the `ThemeData` class, we need to modify the `src/Domain/Entity/ThemeData.php` file.

Here is the *edit block* for this change:

```php
src/Domain/Entity/ThemeData.php
<<<<<<< HEAD
    private ?ThemeData $parentTheme = null;
=======
    public ?ThemeData $parentTheme = null;
>>>>>>> updated
```

This change will make the `$parentTheme` property publicly accessible, allowing it to be accessed directly from an instance of the `ThemeData` class.