aider: Changed type of `$parentTheme` property in `ThemeData` class to `ThemeDataContract`.
# Aider chat conversation:

USER: ThemeData ?ThemeData $parentTheme = null should be ThemeContract

ASSISTANT: To change the type of the `$parentTheme` property in the `ThemeData` class from `?ThemeData` to `?ThemeDataContract`, we need to modify the `src/Domain/Entity/ThemeData.php` file.

Here is the *edit block* for this change:

```php
src/Domain/Entity/ThemeData.php
<<<<<<< HEAD
    public ?ThemeData $parentTheme = null;
=======
    public ?ThemeDataContract $parentTheme = null;
>>>>>>> updated
```

This change will make the `$parentTheme` property accept instances of `ThemeDataContract` or `null`.