aider: Fix issue with setting a theme as active not updating the database record.
# Aider chat conversation:

USER: seems that setting a theme active doesn't update the database record can you check out

ASSISTANT: Sure, I can help with that. However, I need to see the code that is responsible for setting a theme as active. Could you please provide the file that contains this functionality?

USER: seems that setting a theme active doesn't update the database record can you check out from `src/Service/ThemeManager.php` and `src/Repository/ThemeRepository.php`

ASSISTANT: Looking at the `src/Service/ThemeManager.php` and `src/Repository/ThemeRepository.php` files, I can see that the `setActiveTheme` method in the `ThemeManager` service sets the `isActive` property of the `Theme` entity to `true`, but it doesn't seem to persist this change to the database.

To fix this, we need to call the `save` method of the `ThemeRepository` after setting the `isActive` property to `true`. This will ensure that the change is persisted to the database.

Here are the changes we need to make:

In `src/Service/ThemeManager.php`:

```php
src/Service/ThemeManager.php
<<<<<<< HEAD
    public function setActiveTheme(int $id): void
    {
        $theme = $this->findThemeById($id);
        if ($theme) {
            $theme->isActive = true;
            $this->themeRepository->setActive($theme);
        }
    }
=======
    public function setActiveTheme(int $id): void
    {
        $theme = $this->findThemeById($id);
        if ($theme) {
            $theme->isActive = true;
            $this->themeRepository->save($theme);
        }
    }
>>>>>>> updated
```

In `src/Repository/ThemeRepository.php`, we can remove the `setActive` method as it's no longer needed:

```php
src/Repository/ThemeRepository.php
<<<<<<< HEAD
    public function setActive(Theme $theme): void
    {
        $this->save($theme);
    }
=======
>>>>>>> updated
```

These changes should ensure that when a theme is set as active, this change is persisted to the database.