aider: Refactored themes.html.twig to add a link to deactivate a theme if it's currently active.
# Aider chat conversation:

USER: can you check history for QuestionsAnsweredEvent and update the file with latest code version

ASSISTANT: I'm sorry for the confusion, but as an AI, I don't have the ability to access the git history or any other version control system. I can only provide assistance based on the information and files provided to me during our conversation.

If you have the latest version of the `QuestionsAnsweredEvent` file, please provide it and I'll be happy to help with any changes or updates you need.

USER: let's refactor themes.html.twig to be able to deactive a theme too

ASSISTANT: Sure, we can add a new action to the "Actions" column in the table to deactivate a theme. This action will only be displayed if the theme is currently active.

Here are the changes we need to make in `templates/themes.html.twig`:

```twig
templates/themes.html.twig
<<<<<<< HEAD
            <td>
                <a href="{{ path('theme_switch', {themeId: theme.id}) }}">Make Active</a>
                <a href="{{ path('theme_install', {themeId: theme.id}) }}">Install Assets</a>
            </td>
=======
            <td>
                {% if theme.isActive %}
                    <a href="{{ path('theme_deactivate', {themeId: theme.id}) }}">Deactivate</a>
                {% else %}
                    <a href="{{ path('theme_switch', {themeId: theme.id}) }}">Make Active</a>
                {% endif %}
                <a href="{{ path('theme_install', {themeId: theme.id}) }}">Install Assets</a>
            </td>
>>>>>>> updated
```

This change adds a new link to deactivate the theme if it's currently active. The `theme_deactivate` path is a placeholder and should be replaced with the actual path to the action that deactivates the theme. If the theme is not active, the "Make Active" link is displayed as before.

Please note that you will need to implement the corresponding action in your controller to handle the deactivation of a theme.